(#20) Add support for dynamic STOMP destinations.
STOMP destinations can now contain $n references to the nth label in a metric.
So for example, if the destination was "/exchange/graphite/$1" and the metric
was "foo.bar.baz", then the STOMP destination would be
"/exchange/graphite/foo".

Metric batching will only batch while the destination is consistent, if it
changes then any batched metrics are flushed first.